Removing QString conversions from Ruby bindings, use fromStdWString on path->QString conversion, move longPathName to only place it is used
