+++ openstudiocore/src/utilities/core/Path.cpp
@@ -34,30 +34,30 @@
 
 namespace openstudio {
 
-#ifdef Q_OS_WIN
-
-#include <windows.h>
-
-QString longPathName(const QString& path)
-{
-  if (path.isEmpty())
-    return QString();
-  QString maybeShort = QDir::toNativeSeparators(path);
-  QByteArray shortName = maybeShort.toLocal8Bit();
-  char longPath[MAX_PATH];
-  int err = GetLongPathName(shortName.constData(), longPath, MAX_PATH);
-  (void)err;
-  return QDir::fromNativeSeparators(QString::fromLocal8Bit(longPath));
-}
-
-#else
-
-QString longPathName(const QString& path)
-{
-  return path;
-}
-
-#endif
+//#ifdef Q_OS_WIN
+//
+//#include <windows.h>
+//
+//QString longPathName(const QString& path)
+//{
+//  if (path.isEmpty())
+//    return QString();
+//  QString maybeShort = QDir::toNativeSeparators(path);
+//  QByteArray shortName = maybeShort.toLocal8Bit();
+//  char longPath[MAX_PATH];
+//  int err = GetLongPathName(shortName.constData(), longPath, MAX_PATH);
+//  (void)err;
+//  return QDir::fromNativeSeparators(QString::fromLocal8Bit(longPath));
+//}
+//
+//#else
+//
+//QString longPathName(const QString& path)
+//{
+//  return path;
+//}
+//
+//#endif
 
 
 // allow path to be written to cout on Windows
@@ -90,7 +90,7 @@ std::string toString(const path& p)
 /** path to QString. */
 QString toQString(const path& p)
 {
-  return toQString(toString(p));
+  return QString::fromStdWString(p.wstring());
 }
 
 /** UTF-8 encoded char* to path*/
